pyproject.toml
[project]
name = "E-commerce Chatbot"
version = "0.1.0"
description = "An intelligent e-commerce chatbot with product search and FAQ support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "chromadb>=0.5.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "semantic-router>=0.1.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.35.0",
]
